fix dist tests
