--- jest-node-ssr.config.cjs
@@ -1,13 +1,0 @@
-module.exports = {
-  testMatch: ['**/src/**/*.ssr-test.js'],
-  testURL: 'http://localhost',
-  testEnvironment: 'node',
-  transform: {
-    '^.+\\.m?(|j|t)sx?$': [
-      'esbuild-jest-transform',
-      {
-        sourcemap: true,
-      },
-    ],
-  },
-}
